+++ static/js/bootstrap.js
@@ -3,8 +3,35 @@
  */
 export default class bootstrap {
 
-	constructor() {
-		// App Config
-		this.config = window.config;
+	static config() {
+		return window.config;
+	}
+
+	static createCookie(name, value, days) {
+		let expires;
+		if (days) {
+			let date = new Date();
+			date.setTime(date.getTime() + (days * 24 * 60 * 60 * 1000));
+			expires = "; expires=" + date.toGMTString();
+		}
+		else {
+			expires = "";
+		}
+		document.cookie = name + "=" + value + expires + "; path=/";
+	}
+
+	static getCookie(c_name) {
+		if (document.cookie.length > 0) {
+			let c_start = document.cookie.indexOf(c_name + "=");
+			if (c_start != -1) {
+				c_start = c_start + c_name.length + 1;
+				let c_end = document.cookie.indexOf(";", c_start);
+				if (c_end == -1) {
+					c_end = document.cookie.length;
+				}
+				return unescape(document.cookie.substring(c_start, c_end));
+			}
+		}
+		return "";
 	}
 }